Update test.c
Verstable essentially requires a hash function that returns a 64-bit unsigned integer with entropy in the high, as well as low, bits. This is because the four highest bits of each key are stored and used during lookups to skip most failed key comparisons (each of which requires not only calling the comparison function but also accessing the buckets array, i.e. a likely cache miss). Without this mechanism, Verstable can still function relatively well, but it will incur a significant performance hit. The proposed hash function is the same mixer that ankerl::unordered_dense applies internally to guard against weak hash functions.